+++ website/src/components/ToSWrapper.tsx
@@ -1,11 +1,11 @@
-import { Flex, Text } from "@chakra-ui/react";
+import { Box, Flex, Text } from "@chakra-ui/react";
 import { useSession } from "next-auth/react";
 import { useTranslation } from "next-i18next";
 import { ReactNode, useMemo } from "react";
 import { SubmitButton } from "src/components/Buttons/Submit";
 import { SurveyCard } from "src/components/Survey/SurveyCard";
+import { TermsOfService } from "src/components/ToS";
 import { post } from "src/lib/api";
-import { TermsOfService } from "src/pages/terms-of-service";
 
 const navigateAway = () => {
   location.href = "https://laion.ai/";
@@ -24,21 +24,23 @@ export const ToSWrapper = ({ children }: { children?: ReactNode | undefined }) =
 
   const contents = useMemo(
     () => (
-      <SurveyCard display="flex" flexDir="column" w="full" maxWidth="7xl" m="auto" gap={4}>
-        <Text fontWeight="bold" fontSize="xl" as="h2">
-          {t("title")}
-        </Text>
-        <Text>{t("content")}</Text>
-        <TermsOfService />
-        <Flex gap={10} justifyContent="center">
-          <SubmitButton onClick={navigateAway} colorScheme="red">
-            {t("decline")}
-          </SubmitButton>
-          <SubmitButton onClick={acceptToS} colorScheme="blue">
-            {t("accept")}
-          </SubmitButton>
-        </Flex>
-      </SurveyCard>
+      <Box className="oa-basic-theme">
+        <SurveyCard display="flex" flexDir="column" w="full" maxWidth="7xl" m="auto" gap={4}>
+          <Text fontWeight="bold" fontSize="xl" as="h2">
+            {t("title")}
+          </Text>
+          <Text>{t("content")}</Text>
+          <TermsOfService />
+          <Flex gap={10} justifyContent="center">
+            <SubmitButton onClick={navigateAway} colorScheme="red">
+              {t("decline")}
+            </SubmitButton>
+            <SubmitButton onClick={acceptToS} colorScheme="blue">
+              {t("accept")}
+            </SubmitButton>
+          </Flex>
+        </SurveyCard>
+      </Box>
     ),
     [t]
   );
